modify: Adjust table actions of targets and breadcrumbs of municipality

+++ app/Filament/Resources/SkillPriorityResource.php
@@ -89,7 +89,7 @@ class SkillPriorityResource extends Resource
                     ->placeholder('Enter number of slots')
                     ->required()
                     ->markAsRequired(false)
-                    ->integer()
+                    ->numeric()
                     ->disabled(fn($livewire) => $livewire->isEdit())
                     ->dehydrated(),
